Avoid "ready to land" terminology
Report only that a change has had a positive review, allowing the
submitter to follow usual workflow conventions in FreeBSD. We use
Phabricator as a standalone tool to facilitate review if desired,
but its use is not required, and does not dictate our workflow.

Claiming a change is "ready to land" after one positive review does not
mesh with the usual workflow in the FreeBSD community. Some changes may
require more than one reviewer or additional testing, for example.