[Early] Allow differential changesets to be marked as generated via diff properties
Summary:
Ref T784. Allow differential changesets to be marked as generated with a `DifferentialDiffProperty`. The expectation is that `arc diff` will submit an `arc:generated` diff property (which will be populated out of the local `.arcconfig` file). These paths will then be automatically marked as generated when rendering the differential changeset.

Currently, the only way to achieve this is with `differential.generated-paths`. The main issue with `differential.generated-paths` is that it is a global setting and cannot be varied across repositories. A natural extension from `differential.generated-paths` would be to move this configuration into the per-repository configuration, however I feel that this approach would be too inflexible to be useful. Specifically, this setting would need to be updated whenever generated paths are added/removed/changed for a particular repository (although I would expect this to be a fairly infrequent occurrence).

Unfortunately, this functionality will not be retrofitted to existing differentials. We //could// possibly add a migration which populates `arc:generated` for all existing differentials based on the value of `differential.generated-paths`, although I'm not convinced that this is worth pursuing.

Test Plan:
Uploaded a diff (using `arc diff`) and set paths for `arc:generated` with `arc call-conduit 'differential.setdiffproperty'`. Loaded the differential in the UI and noticed that the paths were marked as generated as expected.

One interesting thing to note (I'm not sure if I would consider this to be a bug or expected behavior) is that lint messages takes a higher precedence over generated paths. If a path matches a regular expression in `arc:generated` but has some linter messages associated with that path, then the file will //not// be shown as generated in the UI. This is mainly an issue until after D11537.

Reviewers: epriestley, #blessed_reviewers

Subscribers: Korvin, epriestley

Maniphest Tasks: T784

Differential Revision: https://secure.phabricator.com/D11535