[Early] Don't force main menu search form to use POST
Summary:
This allows browsers (Chorme, Firefox) to add Phabricator as a search engine.

The CSRF token that was previously generated (as it was a POST) was not checked anyway, and I don't think this is exploitable.

Test Plan: Right clicked on Chrome, as "Add As Search Engine". Tried out the search and it worked.

Reviewers: #blessed_reviewers, epriestley

Subscribers: Korvin, epriestley

Differential Revision: https://secure.phabricator.com/D10954